Assignment_4.1
You're going to write a movie storybook by acting in it with your teammates and friends. You're going to take pictures of key scenes in the movie, turn them into cartoons to make the pictures look like drawings, then stage and direct the pictures into the production of the storybook of a movie. You will pick one of the contenders for this year's American, Chinese, or Hindi Oscars (or the movie you like) as the base story and then modify it in a creative way.

You'll be in a team of 2, 3, or 4 students. There will be two submissions per team: i) The storybook as a PDF file, an English version, and a version in your language. The whole story book needs to have at least 5 pages (excluding the poster page), and each page needs to have around 20 pictures. ii) The software for the entire storybook. You are not allowed to stitch pictures together with paintbrush programs; it all needs to be in Python.