Assignment_4.2
Each team submits one zip file, including: i) the storybook as a PDF file, an English version, and a version in your language. The whole story book needs to have at least 5 pages (excluding the poster page), and each page needs to have around 20 pictures. ii) The software for the entire storybook. You are not allowed to stitch pictures together with paintbrush programs, it all needs to be in Python.