Assignment_7
Re-write the p function so that it is now polymorphic with respect to the number of arguments. The last argument is always either a set or a probability distribution dictionary. The other arguments are either sets or predicate functions.

The p function should be able to run like that: 

p(a_1, a_2, a_3, ..., MM)
where the probability is p (a_1| a_2, a_3, ...) operating on either the set or the dictionary MM.